backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.2.0
